added error handling

+++ src/main/java/ca/mcgill/ecse/coolsupplies/controller/CoolSuppliesFeatureSet3Controller.java
@@ -16,30 +16,28 @@ import ca.mcgill.ecse.coolsupplies.application.CoolSuppliesApplication;
 public class CoolSuppliesFeatureSet3Controller {
 
   private static final CoolSupplies coolSupplies = CoolSuppliesApplication.getCoolSupplies();
-  private static List<Item> items;
 
   public static String addItem(String name, int price) {
-    
-    
-    try {
 
-      // Create new item
-      Item new_item = new Item(name, price, coolSupplies);
-      coolSupplies.addItem(new_item);
-    } 
-
-    catch (Exception e) {  // TODO: Fix return strings
-
-      //Check if new_item already exists
-      if (e.getMessage().equals("change this message")) {
-        return "Item already exists";
-      }
-
-      // Other errors (invalid arguments)
-      else {
-        return "Invalid argument(s)";
-      }
+    // Validate Price
+    if (price < 0) {
+      return "The price must be greater than or equal to 0.";
     }
+
+    // Validate name - Name must not be empty
+    if (name.equals("")) {
+      return "The name must not be empty.";
+    }
+
+    // Validate name - name cannot already be in system
+    Item item = (Item) Item.getWithName(name);
+    if (item != null) {
+      return "The name must be unique.";
+    }
+
+    // Add new item - input is valid
+    Item newItem = new Item(name, price, coolSupplies);
+    coolSupplies.addItem(newItem);
 
     return "";  // Success message
     
@@ -47,11 +45,33 @@ public class CoolSuppliesFeatureSet3Controller {
 
   public static String updateItem(String name, String newName, int newPrice) {
     
-    // TODO: Handle errors
-    // Get Item then set new name and new price
+
+    // Validate input - item must exist
     Item item = (Item) Item.getWithName(name);
+    if (item == null) {
+      return "The item does not exist.";
+    }
+
+    // Validate input - newName must not be empty.
+    if (newName.equals("")) {
+      return "The name must not be empty.";
+    }
+
+    // Validate Price
+    if (newPrice < 0) {
+      return "The price must be greater than or equal to 0.";
+    }
+
+    // Validate newName - Make sure newName is not already in system
+    Item newNameItem = (Item) Item.getWithName(newName);
+    if (newNameItem != null) {
+      return "The name must be unique.";
+    }
+
+    // Set new name and new price
     item.setName(newName);
     item.setPrice(newPrice);
+    
     return "";
   }
 
@@ -60,13 +80,19 @@ public class CoolSuppliesFeatureSet3Controller {
     // Check if list of items is empty
     if (coolSupplies.hasItems()) {
 
-      // If list is not empty, get item and delete it
       Item item = (Item) Item.getWithName(name);
+
+      // Check if item exists
+      if (item == null) {
+        return "The item does not exist.";
+      }
+
+      // If list is not empty, get item and delete it
       item.delete();
 
       return "";
     }
-    else return "Item list is empty";
+    else return "";
   }
 
   public static TOItem getItem(String name) {
